backend/pyproject.toml
[project]
name = "tozo"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "d1mnaz", email = "dimonnaz@gmail.com"},
]
dependencies = [
    "quart>=0.20.0",
    "quart-auth>=0.11.0",
    "argon2-cffi>=25.1.0",
    "zxcvbn>=4.5.0",
    "itsdangerous>=2.2.0",
    "quart-rate-limiter>=0.12.1",
    "pydantic[email]>=2.12.5",
    "quart-schema>=0.23.0",
    "quart-db>=0.11.0"
]
requires-python = "==3.14.*"
readme = "README.md"
license = {text = "MIT"}


[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "freezegun>=1.5.5",
]

[tool.pdm.scripts]
start = {cmd = "quart --app src/backend/run.py run", env_file = "development.env"}
# test = {cmd = "pytest tests/", env_file = "testing.env"}
recreate-db-base = "quart --app src/backend/run.py recreate_db"
recreate-db = {composite = ["recreate-db-base"], env_file =  "development.env"}
test = {composite = ["recreate-db-base", "pytest tests/"], env_file = "testing.env"}
lint = "ruff check src tests"
format = "ruff check --fix  src tests"


[tool.pytest.ini_options]
addopts = "-v --no-header --showlocals"
asyncio_mode = "auto"
pythonpath = [
    "src",
]

[tool.ruff.lint]
select = [
    "UP",   # pyupgrade
    "I",    # isort
    "E",    # flake8-errors
    "W",    # flake8-warnings
    "F",    # flake8-pyflakes
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "S",    # flake8-bandit
    "BLE",  # flake8-blind-except
    "T20",  # flake8-print
    "T10",  # flake8-debugger
    "PT",   # flake8-pytest-style
    "RUF",  # Ruff
    "DTZ",  # flake8-datetimez
    "ERA",  # eradicate
    "FURB", # refurb
]

ignore = [
    "C408",  # Ненужный вызов dict, list и т.д.
    "C416",  # Ненужный вызов dict, list и т.д.
    "RUF001", # Игнорирование похожих символов с разными ASCII
    "RUF002", # Игнорирование похожих символов с разными ASCII
    "RUF003", # Игнорирование похожих символов с разными ASCII
    "S101",  # assert
    "S603",
    "S607",
    "S110",  # try-except-pass
    "S311",  # pseudo-random generators
    "E501",
    "ERA001",
]

fixable = [
    "I", # cортировка импортов
]

[tool.pdm]
distribution = false
